run SCION as services on Windows
